+++ Lab4/exercise2/Colour.java
