+++ Lab4/exercise2/Point.java
